Only update repository if branch is missing.

+++ .ci/post-release-repo-branches.sh
@@ -43,38 +43,45 @@ previous_version="${previous_version:1:${#previous_version}}" # remove 'v' prefi
 # branch and create a new release branch.
 release_branch="release/${previous_version}"
 release_minor_branch="release/${previous_version%.*}.x"
-echo "Creating a new ${release_branch} branch..."
 
-wrap git checkout stable-website \
-     "Failed to checkout stable-website branch"
-wrap git checkout -b "${release_branch}" \
-     "Failed to create new branch: ${release_branch}"
-wrap git push origin "${release_branch}" \
-     "Failed to push new branch: ${release_branch}"
+# If the release branch already exists, skip this so we don't muck
+# up the various branches
+if ! wrap_raw git ls-remote --exit-code --heads origin "${release_branch}"; then
+    echo "Creating a new ${release_branch} branch..."
 
-echo "Creating a new ${release_minor_branch} branch..."
-wrap git checkout stable-website \
-     "Failed to checkout stable-website branch"
-# The release minor branch will likely not exist locally but
-# just be sure it's not there
-git branch -d "${release_minor_branch}" > "${output}" 2>&1
+    wrap git checkout stable-website \
+        "Failed to checkout stable-website branch"
+    wrap git checkout -b "${release_branch}" \
+        "Failed to create new branch: ${release_branch}"
+    wrap git push origin "${release_branch}" \
+        "Failed to push new branch: ${release_branch}"
 
-wrap git checkout -b "${release_minor_branch}" \
-     "Failed to create new branch: ${release_minor_branch}"
-wrap git push -f origin "${release_minor_branch}" \
-     "Failed to push new branch: ${release_minor_branch}"
+    echo "Creating a new ${release_minor_branch} branch..."
+    wrap git checkout stable-website \
+        "Failed to checkout stable-website branch"
+    # The release minor branch will likely not exist locally but
+    # just be sure it's not there
+    git branch -d "${release_minor_branch}" > "${output}" 2>&1
 
-slack -m "New branches created for previous release: ${release_branch} and ${release_minor_branch}"
+    wrap git checkout -b "${release_minor_branch}" \
+        "Failed to create new branch: ${release_minor_branch}"
+    wrap git push -f origin "${release_minor_branch}" \
+        "Failed to push new branch: ${release_minor_branch}"
 
-# Now update the stable-website branch with new version content
-echo "Updating stable-website with latest..."
-wrap git checkout "v${vagrant_version}" \
-     "Failed to checkout vagrant release v${vagrant_version}"
-wrap git branch -d stable-website \
-    "Failed to delete the stable-website branch"
-wrap git checkout -b stable-website \
-     "Failed to create new stable-website branch from release v${vagrant_version}"
-wrap git push -f origin stable-website \
-     "Failed to push updated stable-website branch"
+    slack -m "New branches created for previous release: ${release_branch} and ${release_minor_branch}"
 
-slack -m "Vagrant website changes have been pushed for release ${vagrant_version}"
+    # Now update the stable-website branch with new version content
+    echo "Updating stable-website with latest..."
+    wrap git checkout "v${vagrant_version}" \
+        "Failed to checkout vagrant release v${vagrant_version}"
+    wrap git branch -d stable-website \
+        "Failed to delete the stable-website branch"
+    wrap git checkout -b stable-website \
+        "Failed to create new stable-website branch from release v${vagrant_version}"
+    wrap git push -f origin stable-website \
+        "Failed to push updated stable-website branch"
+
+    slack -m "Vagrant website changes have been pushed for release ${vagrant_version}"
+else
+    echo "Skipping repository updates because branch ${release_branch} already exists"
+fi
